+++ demis/applications/outputs.tf
@@ -30,32 +30,32 @@ output "spp_ars_enabled" {
 
 output "fhir_profile_snapshots" {
   description = "Version of the FHIR Profile Snapshots being used"
-  value       = local.vs_core_enabled && length(tolist(module.validation_service_core_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_core_metadata[0].current_profile_versions))}]" : local.fhir_profile_snapshots
+  value       = local.vs_core_enabled ? "[${join(", ", tolist(module.validation_service_core_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "igs_profile_snapshots" {
   description = "Version of the IGS Profile Snapshots being used"
-  value       = local.vs_igs_enabled && length(tolist(module.validation_service_igs_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_igs_metadata[0].current_profile_versions))}]" : local.igs_profile_snapshots
+  value       = local.vs_igs_enabled ? "[${join(", ", tolist(module.validation_service_igs_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "ars_profile_snapshots" {
   description = "Version of the ARS Profile Snapshots being used"
-  value       = local.vs_ars_enabled && length(tolist(module.validation_service_ars_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_ars_metadata[0].current_profile_versions))}]" : local.ars_profile_snapshots
+  value       = local.vs_ars_enabled ? "[${join(", ", tolist(module.validation_service_ars_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "statistic_package_versions" {
   description = "Version of the statistic package being used"
-  value       = local.vs_bedoccupancy_enabled && length(tolist(module.validation_service_bedoccupancy_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_bedoccupancy_metadata[0].current_profile_versions))}]" : local.fhir_profile_snapshots
+  value       = local.vs_bedoccupancy_enabled ? "[${join(", ", tolist(module.validation_service_bedoccupancy_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "disease_package_versions" {
   description = "Version of the disease package being used"
-  value       = local.vs_disease_enabled && length(tolist(module.validation_service_disease_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_disease_metadata[0].current_profile_versions))}]" : local.fhir_profile_snapshots
+  value       = local.vs_disease_enabled ? "[${join(", ", tolist(module.validation_service_disease_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "laboratory_package_versions" {
   description = "Version of the statistic package being used"
-  value       = local.vs_pathogen_enabled && length(tolist(module.validation_service_pathogen_metadata[0].current_profile_versions)) > 0 ? "[${join(", ", tolist(module.validation_service_pathogen_metadata[0].current_profile_versions))}]" : local.fhir_profile_snapshots
+  value       = local.vs_pathogen_enabled ? "[${join(", ", tolist(module.validation_service_pathogen_apps[0].profile_metadata.current_profile_versions))}]" : "[]"
 }
 
 output "version_istio_routing_chart" {