+++ modules/validation_service_apps/outputs.tf
@@ -1,0 +1,4 @@
+output "profile_metadata" {
+  description = "The metadata of the FHIR profiles used for the Validation Service, including their versions and the destination subsets for Istio routing."
+  value       = module.validation_service_metadata
+}